% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lfwtr}
\alias{lfwtr}
\title{Recursively List Files With Type}
\usage{
lfwtr(dir, ...)
}
\arguments{
\item{dir}{a character string with the path name to a directory.}

\item{\dots}{Additional arguments, passed to \code{\link{list_files_with_type}}.}
}
\description{
List files with type with recursive support
}
\keyword{internal}
