% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDiff-package.R
\docType{package}
\name{packageDiff-package}
\alias{packageDiff}
\alias{packageDiff-package}
\title{Package Diff}
\description{
Extract package information and compare between versions.
}
\author{
Cole Beck \email{cole.beck@vumc.org}

Maintainer: Cole Beck \email{cole.beck@vumc.org}
}
