% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiff.R
\name{pkgDiff}
\alias{pkgDiff}
\title{Package Diff}
\usage{
pkgDiff(a, b, doc = TRUE, src = TRUE, vignettes = TRUE,
  tests = TRUE, demo = TRUE, width = 80)
}
\arguments{
\item{a}{First package, a \sQuote{pkgInfo} object.}

\item{b}{Second package, a \sQuote{pkgInfo} object.}

\item{doc}{Include documentation in diff output.}

\item{src}{Include src files in diff output}

\item{vignettes}{Include vignette files in diff output}

\item{tests}{Include test files in diff output}

\item{demo}{Include demo files in diff output}

\item{width}{Output width.}
}
\value{
\sQuote{diffr} object is open in browser
}
\description{
This function compares version changes within packages.
}
\details{
Generate diffs between package information.
}
\examples{
tar1 <- system.file("examples", "yaml_2.1.18.tar.gz", package = "packageDiff")
tar2 <- system.file("examples", "yaml_2.1.19.tar.gz", package = "packageDiff")
a <- pkgInfo(tar1)
b <- pkgInfo(tar2)
pkgDiff(a, b)

}
