% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.packageInfo}
\alias{print.packageInfo}
\title{Print Package Information}
\usage{
\method{print}{packageInfo}(x, ...)
}
\arguments{
\item{x}{A packageInfo object.}

\item{\dots}{Additional parameters, unused at this time.}
}
\description{
Print method for \sQuote{packageInfo} class.
}
\details{
Print its argument and return it invisibly.
}
\examples{
\dontrun{
url <- 'https://cran.r-project.org/src/contrib/Archive/acepack/acepack_1.3-3.3.tar.gz'
info <- packageInfo(url)
print(info)
}
}
