% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.pkgInfo}
\alias{print.pkgInfo}
\title{Print Package Information}
\usage{
\method{print}{pkgInfo}(x, doc = FALSE, src = FALSE,
  vignettes = FALSE, tests = FALSE, demo = FALSE, ...)
}
\arguments{
\item{x}{A packageInfo object.}

\item{doc}{Include documentation in output.}

\item{src}{Include src files in output.}

\item{vignettes}{Include vignette files in output.}

\item{tests}{Include test files in output.}

\item{demo}{Include demo files in output.}

\item{\dots}{Additional parameters, unused at this time.}
}
\description{
Print method for \sQuote{packageInfo} class.
}
\details{
Print its argument and return it invisibly.
}
\examples{
tarfile <- system.file("examples", "acepack_1.3-3.3.tar.gz", package = "packageDiff")
pkgInfo(tarfile)
\donttest{
url <- 'https://cran.r-project.org/src/contrib/Archive/acepack/acepack_1.3-3.3.tar.gz'
info <- pkgInfo(url)
print(info)
}
}
