% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scanner}
\alias{scanner}
\title{Scan List of Text Files}
\usage{
scanner(files, full.names = FALSE)
}
\arguments{
\item{file}{the names of files to read.}
}
\description{
Store text file as character string
}
\keyword{internal}
