% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setOrReplace}
\alias{setOrReplace}
\alias{\%!\%}
\title{Set or Replace}
\usage{
lhs \%!\% rhs
}
\arguments{
\item{lhs}{thing to check}

\item{rhs}{thing to replace}
}
\description{
Set or replace empty
}
\keyword{internal}
