% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sourcerer}
\alias{sourcerer}
\title{Source Many R Scripts}
\usage{
sourcerer(files, envir = NULL)
}
\arguments{
\item{files}{Vector of file names.}

\item{envir}{Default environment.}
}
\description{
Evaluate R scripts in protected environment.
}
\details{
This may fail if Collate is incorrect.
}
