% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipPackage.R
\name{unzipPackage}
\alias{unzipPackage}
\title{Extract R Package}
\usage{
unzipPackage(x)
}
\arguments{
\item{x}{The compressed (tar.gz) build file of an R package, either local or
URL.}
}
\value{
List of files extracted.
}
\description{
Untar an R package into a temp directory.
}
