% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDiff-package.R
\docType{package}
\name{packageDiff-package}
\alias{packageDiff}
\alias{packageDiff-package}
\title{Package Diff}
\description{
Extract package information and compare between versions.
}
\details{
This package is experimental. Please submit bugs to
\url{https://github.com/couthcommander/packageDiff}.
}
\examples{
\dontrun{
a <- pkgInfo('https://cran.r-project.org/src/contrib/Archive/yaml/yaml_2.1.18.tar.gz')
b <- pkgInfo('https://cran.r-project.org/src/contrib/Archive/yaml/yaml_2.1.19.tar.gz')
pkgDiff(a, b)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/couthcommander/packageDiff}
}

}
\author{
Cole Beck \email{cole.beck@vumc.org}

Maintainer: Cole Beck \email{cole.beck@vumc.org}
}
