% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDiff.R
\name{pkgDiff}
\alias{pkgDiff}
\title{Package Diff}
\usage{
pkgDiff(a, b, doc = TRUE, width = 200)
}
\arguments{
\item{a}{First package, a \sQuote{pkgInfo} object.}

\item{b}{Second package, a \sQuote{pkgInfo} object.}

\item{doc}{Include documentation in diff output.}

\item{width}{Output width.}
}
\value{
\sQuote{diffr} object is open in browser
}
\description{
This function compares version changes within packages.
}
\details{
Generate diffs between package information.
}
